backend/requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3.0

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Task Queue
celery==5.3.4

# Environment Variables
python-dotenv==1.0.0

# Development Tools
black==23.10.1
flake8==6.1.0
pytest==7.4.3
pytest-django==4.5.2
coverage==7.3.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
pytz==2023.3.post1
Pillow==10.1.0
